Add one more test case and split into new files
