+++ packages/versioning/src/versioning.ts
@@ -1,5 +1,4 @@
 import {
-  compilerAssert,
   DecoratorContext,
   DiagnosticTarget,
   Enum,
@@ -14,6 +13,7 @@ import {
   Type,
 } from "@typespec/compiler";
 import { createStateSymbol, reportDiagnostic } from "./lib.js";
+import { TimelineMoment, VersioningTimeline } from "./versioning-timeline.js";
 
 const addedOnKey = createStateSymbol("addedOn");
 const removedOnKey = createStateSymbol("removedOn");
@@ -538,125 +538,6 @@ export function resolveVersions(program: Program, rootNs: Namespace): VersionRes
 }
 
 /**
- * Represent a timeline of all the version involved in the versioning of a namespace
- * Starting at 0.
- */
-export class VersioningTimeline {
-  #namespaces: Namespace[];
-  #timeline: TimelineMoment[];
-  #versionIndex: Map<Version, number>;
-
-  constructor(program: Program, resolutions: Map<Namespace, Version>[]) {
-    const indexedVersions = new Set<Version>();
-    const namespaces = new Set<Namespace>();
-    const timeline = resolutions.map((x) => new TimelineMoment(x));
-    for (const resolution of resolutions) {
-      for (const [namespace, version] of resolution.entries()) {
-        indexedVersions.add(version);
-        namespaces.add(namespace);
-      }
-    }
-
-    for (const namespace of namespaces) {
-      const [, versions] = getVersions(program, namespace);
-      if (versions === undefined) {
-        continue;
-      }
-
-      for (const version of versions.getVersions()) {
-        if (!indexedVersions.has(version)) {
-          indexedVersions.add(version);
-
-          timeline.push(new TimelineMoment(new Map([[version.namespace, version]])));
-        }
-      }
-    }
-
-    // Order the timeline
-
-    for (const namespace of namespaces) {
-      timeline.sort((a, b) => {
-        const aVersion = a.getVersion(namespace);
-        const bVersion = b.getVersion(namespace);
-
-        return aVersion === undefined || bVersion === undefined
-          ? 0
-          : aVersion.index - bVersion.index;
-      });
-    }
-    this.#timeline = timeline;
-    this.#namespaces = [...namespaces];
-
-    this.#versionIndex = new Map();
-    for (const [index, moment] of timeline.entries()) {
-      for (const version of moment.versions()) {
-        this.#versionIndex.set(version, index);
-      }
-    }
-  }
-
-  prettySerialize() {
-    const hSep = "-".repeat(this.#namespaces.length * 13 + 1);
-    const content = this.#timeline
-      .map((moment) => {
-        return (
-          "| " +
-          this.#namespaces
-            .map((x) => (moment.getVersion(x)?.name ?? "").padEnd(10, " "))
-            .join(" | ") +
-          " |"
-        );
-      })
-      .join(`\n${hSep}\n`);
-    return ["", hSep, content, hSep].join("\n");
-  }
-
-  get(version: Version): TimelineMoment {
-    const index = this.getIndex(version);
-    return this.#timeline[index];
-  }
-
-  /**
-   * Return index in the timeline that this version points to
-   */
-  getIndex(version: Version): number {
-    const index = this.#versionIndex.get(version);
-    compilerAssert(
-      index !== undefined,
-      `Version "${version?.name}" from ${version.namespace.name}  should have been resolved`
-    );
-    return index;
-  }
-
-  first(): TimelineMoment {
-    return this.#timeline[0];
-  }
-
-  [Symbol.iterator](): IterableIterator<TimelineMoment> {
-    return this.#timeline[Symbol.iterator]();
-  }
-  entries(): IterableIterator<[number, TimelineMoment]> {
-    return this.#timeline.entries();
-  }
-}
-
-export class TimelineMoment {
-  #versionMap: Map<Namespace, Version>;
-
-  public constructor(versionMap: Map<Namespace, Version>) {
-    this.#versionMap = versionMap;
-  }
-
-  getVersion(namespace: Namespace): Version | undefined {
-    return this.#versionMap.get(namespace);
-  }
-
-  versions(): IterableIterator<Version> {
-    return this.#versionMap.values();
-  }
-}
-
-/**
  * Represent the set of projections used to project to that version.
  */
 interface VersionProjections {
@@ -707,7 +588,6 @@ export function buildVersionProjections(program: Program, rootNs: Namespace): Ve
     program,
     resolutions.map((x) => x.versions)
   );
-  console.log("Timeline", timeline.prettySerialize());
   return resolutions.map((resolution) => {
     if (resolution.versions.size === 0) {
       return { version: undefined, projections: [] };
@@ -811,12 +691,6 @@ export enum Availability {
   Removed = "Removed",
 }
 
-/**
- * Returns a map of version names and whether a given type is available in that version
- * @param program TypeSpec program
- * @param type Type to get the availability map for
- * @returns the availability map for the type, if applicable
- */
 export function getAvailabilityMap(
   program: Program,
   type: Type
@@ -886,25 +760,8 @@ export function getAvailabilityMapV2(
   // something isn't available by default
   let isAvail = false;
   for (const [index, moment] of timeline.entries()) {
-    // const add = added.find((x) => x === moment.getVersion(getOriginalNamespace(x.namespace)));
-    // const rem = removed.find((x) => x === moment.getVersion(getOriginalNamespace(x.namespace)));
-
-    console.log(
-      "Add is in projection",
-      added.map((x) => x.enumMember.projectionSource !== undefined)
-    );
     const add = added.find((x) => timeline.getIndex(x) === index);
     const rem = removed.find((x) => timeline.getIndex(x) === index);
-    if ((type as any).name === "b") {
-      console.log(
-        "Check ver set",
-        "Added at:",
-        added.map((x) => x.name),
-        "Get version moment",
-        add?.name,
-        rem?.name
-      );
-    }
     if (rem) {
       isAvail = false;
       avail.set(moment, Availability.Removed);
@@ -927,11 +784,6 @@ export function existsAtVersion(p: Program, type: Type, versionKey: ObjectType):
   const availability = getAvailabilityMapV2(p, type, versioningState.timeline);
   if (!availability) return true;
   const isAvail = availability.get(versioningState.projectingMoment)!;
-
-  if ((type as any).name === "b") {
-    isAvail === Availability.Added || isAvail === Availability.Available;
-  }
-
   return isAvail === Availability.Added || isAvail === Availability.Available;
 }
